+++ plugins/destination/motherduck/client/client_test.go
@@ -43,6 +43,9 @@ func TestPlugin(t *testing.T) {
 				AddColumn:    true,
 				RemoveColumn: true,
 			},
+			SkipSpecificMigrations: plugin.Migrations{
+				RemoveUniqueConstraint: true,
+			},
 		},
 		plugin.WithTestDataOptions(schema.TestSourceOptions{
 			// not supported in Parquet Writer