+++ scaffold/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [3.0.7](https://github.com/cloudquery/cloudquery/compare/scaffold-v3.0.6...scaffold-v3.0.7) (2024-03-18)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-pb-go to v1.19.3 ([#17219](https://github.com/cloudquery/cloudquery/issues/17219)) ([b5e5c19](https://github.com/cloudquery/cloudquery/commit/b5e5c190a47191e3792843ba3b1d7b0cd5d6522f))
+* **deps:** Update module github.com/cloudquery/plugin-pb-go to v1.19.4 ([#17228](https://github.com/cloudquery/cloudquery/issues/17228)) ([8a051f3](https://github.com/cloudquery/cloudquery/commit/8a051f36f57b99f92d873a22d2e5f72c36a8e32c))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.34.0 ([#17203](https://github.com/cloudquery/cloudquery/issues/17203)) ([4b128b6](https://github.com/cloudquery/cloudquery/commit/4b128b6722dea883d66458f2f3c831184926353d))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.34.1 ([#17220](https://github.com/cloudquery/cloudquery/issues/17220)) ([08d4950](https://github.com/cloudquery/cloudquery/commit/08d49504aee10f6883e1bd4f7e1102a274c8ee81))
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.34.2 ([#17229](https://github.com/cloudquery/cloudquery/issues/17229)) ([41ed721](https://github.com/cloudquery/cloudquery/commit/41ed721cfa435a4937f3022501dd4d45a3a880b0))
+
+## [3.0.6](https://github.com/cloudquery/cloudquery/compare/scaffold-v3.0.5...scaffold-v3.0.6) (2024-03-12)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.32.1 ([#17044](https://github.com/cloudquery/cloudquery/issues/17044)) ([d3592e7](https://github.com/cloudquery/cloudquery/commit/d3592e7f3ae600655778eb508aeccfa4e5b74e8c))
+
 ## [3.0.5](https://github.com/cloudquery/cloudquery/compare/scaffold-v3.0.4...scaffold-v3.0.5) (2024-03-05)